ctfs/p4ctf/2023/Quals/ai/GeoHacker/app/geotask/requirements.txt
#torchvision
#torchaudio
pygeohash
geopandas
wandb
matplotlib
numpy
torchsummary
scikit-learn
Pillow
haversine
s2cell
geopandas
#cartopy
shapely
